photogallery/requirements.txt
asgiref==3.9.1
beautifulsoup4==4.13.4
certifi==2025.7.9
charset-normalizer==3.4.2
cloudinary==1.44.1
contourpy==1.3.2
crispy-tailwind==1.0.3
cycler==0.12.1
Django==5.2.4
django-crispy-forms==2.4
dnspython==1.16.0
fonttools==4.58.5
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.3.1
packaging==25.0
pandas==2.3.1
pillow==11.3.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pymongo==3.12.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.9.0
